Introduce maps/io.py
